refactor(errors): move isNativeError function
