+++ packages/framework/src/errors/base.errors.ts
@@ -1,6 +1,24 @@
 import { HttpStatusEnum } from '../constants';
 import { ErrorCodeEnum } from '../constants/error.constants';
-import { isNativeError } from './guard.errors';
+
+/**
+ * Check if the object is a native error.
+ *
+ * This method relies on `Object.prototype.toString()` behavior. It is possible to obtain
+ * an incorrect result when the object argument manipulates the `@@toStringTag` property.
+ *
+ * @param object - The object to check.
+ * @returns `true` if the object is a native error, `false` otherwise.
+ */
+export const isNativeError = (object: unknown): object is Error => {
+  if (typeof object !== 'object' || object === null) {
+    return false;
+  }
+
+  const proto = Object.getPrototypeOf(object);
+
+  return proto?.constructor?.name.endsWith('Error') ?? false;
+};
 
 /**
  * Base error class.
