+++ packages/framework/src/errors/guard.errors.ts
@@ -3,25 +3,6 @@ import { FrameworkError } from './base.errors';
 import { PlatformError } from './platform.errors';
 
 /**
- * Check if the object is a native error.
- *
- * This method relies on `Object.prototype.toString()` behavior. It is possible to obtain
- * an incorrect result when the object argument manipulates the `@@toStringTag` property.
- *
- * @param object - The object to check.
- * @returns `true` if the object is a native error, `false` otherwise.
- */
-export const isNativeError = (object: unknown): object is Error => {
-  if (typeof object !== 'object' || object === null) {
-    return false;
-  }
-
-  const proto = Object.getPrototypeOf(object);
-
-  return proto?.constructor?.name.endsWith('Error') ?? false;
-};
-
-/**
  * Check if the error is a `FrameworkError`.
  *
  * A `FrameworkError` is an error thrown by the Framework.
